requirements_txt.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
openai==0.28.1
pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
python-multipart==0.0.6